+++ Assets/Tests/InputSystem/Plugins/EnhancedTouchTests.cs
@@ -139,12 +139,13 @@ internal class EnhancedTouchTests : CoreTestsFixture
     [TestCase(InputSettings.UpdateMode.ProcessEventsInFixedUpdate)]
     public void EnhancedTouch_SupportsEditorUpdates(InputSettings.UpdateMode updateMode)
     {
+        InputSystem.settings.gameViewFocus = default;
+
         // To better observe that play mode and edit mode state is indeed independent and handled
         // correctly, suppress resetting of the touch device when focus is lost to the player.
         runtime.runInBackground = true;
         SetCanRunInBackground(Touchscreen.current);
 
-        InputEditorUserSettings.lockInputToGameView = false;
         InputSystem.settings.updateMode = updateMode;
         runtime.currentTimeForFixedUpdate += Time.fixedDeltaTime;
         // Run one player update with data.